Move Constants to util package
